닷지 무적시간 구현

+++ Source/CyperArena/CustomData.h
@@ -106,7 +106,7 @@ public:
 	UPROPERTY(EditAnyWhere, BlueprintReadWrite)
 		EAttackType attack_type;
 	UPROPERTY(EditAnyWhere, BlueprintReadWrite)
-		uint8 durability_level;
+		float durability_level;
 	UPROPERTY(EditAnyWhere, BlueprintReadWrite)
 		FVector knock_back_offset;
